refactor!: typescript ruler no more
