+++ package.json
@@ -10,15 +10,15 @@
   },
   "dependencies": {
     "@custom-react-hooks/all": "^1.4.19",
-    "@nextui-org/react": "^2.2.10",
-    "formik": "^2.4.5",
+    "@nextui-org/react": "^2.3.0",
     "axios": "^1.6.8",
-    "framer-motion": "^11.0.28",
+    "formik": "^2.4.5",
+    "framer-motion": "^11.1.1",
     "jotai": "^2.8.0",
     "leaflet": "^1.9.4",
     "next": "14.2.1",
     "react": "^18",
-    "react-big-calendar": "^1.11.3",
+    "react-big-calendar": "^1.11.6",
     "react-dom": "^18",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.1.0",
@@ -30,10 +30,10 @@
   },
   "devDependencies": {
     "@biomejs/biome": "1.7.0",
-    "@types/node": "20.12.7",
-    "@types/react": "18.2.78",
+    "@types/node": "^20.12.7",
+    "@types/react": "^18.2.79",
     "@types/react-big-calendar": "^1.8.9",
-    "@types/react-dom": "18.2.25",
+    "@types/react-dom": "^18.2.25",
     "autoprefixer": "^10.4.19",
     "eslint": "^8",
     "eslint-config-next": "14.2.1",
