--- src/features/ui/components/CustomAlert/index.js
@@ -1,1 +1,0 @@
-export { default as CustomAlert } from "./CustomAlert";
