--- src/features/ui/components/CustomInput/CustomInput.d.ts
@@ -1,10 +1,0 @@
-import { InputProps } from "@nextui-org/react";
-import { FC } from "react";
-
-interface Props extends Omit<InputProps, "color" | "variant"> {
-  variant?: "flat" | "bordered";
-}
-
-const CustomInput: FC<Props>;
-
-export default CustomInput;
