--- src/features/ui/components/CustomInput/index.js
@@ -1,1 +1,0 @@
-export { default as CustomInput } from "./CustomInput";
