--- src/features/ui/components/CustomLogo/CustomLogo.d.ts
@@ -1,9 +1,0 @@
-import { FC, ImgHTMLAttributes } from "react";
-
-interface Props extends ImgHTMLAttributes<HTMLImageElement> {
-  color?: "light" | "dark";
-}
-
-const CustomLogo: FC<Props>;
-
-export default CustomLogo;
