--- src/features/ui/components/CustomLogo/index.js
@@ -1,1 +1,0 @@
-export { default as CustomLogo } from "./CustomLogo";
