+++ src/features/ui/components/CustomTable/CustomTableBody.jsx
@@ -1,3 +1,17 @@
+// @ts-check
+
+/**
+ * @param {{
+ *   items: {
+ *     key: number;
+ *     name: string;
+ *     price: number;
+ *     img: string;
+ *     ticket: string;
+ *   }[]
+ * }} props
+ * @returns {React.ReactNode}
+ */
 export default function CustomTableBody({ items }) {
   return (
     <tbody className="whitespace-nowrap text-center">
@@ -7,7 +21,6 @@ export default function CustomTableBody({ items }) {
           className="odd:bg-[#D1E7F1] even:bg-[#E7F0F4] border-t-4 border-white"
         >
           <th className="px-8 py-3 font-medium flex flex-row items-center gap-0 min-h-fit">
-            {/* eslint-disable-next-line */}
             <img
               src={item.img}
               alt={item.name}
