--- src/features/ui/components/CustomTable/CustomTableHeader.d.ts
@@ -1,7 +1,0 @@
-interface Props {
-  headers: string[];
-}
-
-function CustomTableHeader(props: Props): JSX.Element;
-
-export default CustomTableHeader;
