--- src/features/ui/components/CustomTextarea/index.js
@@ -1,1 +1,0 @@
-export { default as CustomTextarea } from "./CustomTextarea";
