+++ src/features/ui/components/index.js
@@ -1,7 +1,8 @@
-export { default as CustomButton } from "./CustomButton";
-export * from "./CustomTable";
-export * from "./CustomInput";
-export * from "./CustomTextarea";
 export * from "./CustomModal";
-export * from "./CustomLogo";
-export * from "./CustomAlert";
+export * from "./CustomTable";
+
+export { default as CustomAlert } from "./CustomAlert";
+export { default as CustomButton } from "./CustomButton";
+export { default as CustomInput } from "./CustomInput";
+export { default as CustomLogo } from "./CustomLogo";
+export { default as CustomTextarea } from "./CustomTextarea";
